Move column hooks into their own files.
