+++ src/features/import/components/Importer/Configure/Configuration/index.tsx
@@ -2,10 +2,11 @@ import { CompareArrows } from '@mui/icons-material';
 import { FC } from 'react';
 import { Box, Divider, Typography, useTheme } from '@mui/material';
 
+import { ConfiguringData } from '..';
 import messageIds from 'features/import/l10n/messageIds';
 import TagConfigRow from './TagConfigRow';
+import useColumn from 'features/import/hooks/useColumn';
 import ZUIEmptyState from 'zui/ZUIEmptyState';
-import { ConfiguringData, useColumn } from '..';
 import {
   ExperimentalFieldTypes,
   ExperimentColumn,
