+++ src/features/import/components/Importer/Configure/index.tsx
@@ -2,11 +2,9 @@ import { Box } from '@mui/material';
 import { FC, useState } from 'react';
 
 import Configuration from './Configuration';
-import messageIds from 'features/import/l10n/messageIds';
-import range from 'utils/range';
-import { useMessages } from 'core/i18n';
-import { ExperimentalFieldTypes, ExperimentColumn } from './Mapping/MappingRow';
-import Mapping, { ExperimentRow } from './Mapping';
+import { ExperimentalFieldTypes } from './Mapping/MappingRow';
+import Mapping from './Mapping';
+import useColumns from 'features/import/hooks/useColumns';
 import SheetSettings, { ExperimentSheet } from './SheetSettings';
 
 export interface ConfiguringData {
@@ -17,56 +15,6 @@ export interface ConfiguringData {
 interface ConfigureProps {
   sheets: ExperimentSheet[];
 }
-
-export const useColumn = (column: (string | number | null)[]) => {
-  const rowsWithValues: (string | number)[] = [];
-  let numberOfEmptyRows = 0;
-
-  column.forEach((rowValue) => {
-    if (typeof rowValue === 'string' || typeof rowValue === 'number') {
-      rowsWithValues.push(rowValue);
-    } else {
-      numberOfEmptyRows += 1;
-    }
-  });
-
-  const uniqueValues = Array.from(new Set(rowsWithValues));
-
-  return { numberOfEmptyRows, uniqueValues };
-};
-
-const useColumns = (
-  firstRowIsHeaders: boolean,
-  rows: ExperimentRow[]
-): ExperimentColumn[] => {
-  const messages = useMessages(messageIds);
-  const numberOfColumns = rows ? rows[0].data.length : 0;
-
-  const columns: ExperimentColumn[] = [];
-  range(numberOfColumns).forEach((number) =>
-    columns.push({ data: [], id: number + 1, title: '' })
-  );
-
-  rows?.forEach((row, rowIndex) => {
-    row.data.forEach((cellValue, cellIndex) => {
-      const column = columns[cellIndex];
-      if (rowIndex == 0) {
-        if (firstRowIsHeaders && cellValue !== null) {
-          column.title = cellValue as string;
-        } else {
-          column.title = messages.configuration.mapping.defaultColumnHeader({
-            columnIndex: cellIndex + 1,
-          });
-          column.data.push(cellValue);
-        }
-      } else {
-        column.data.push(cellValue);
-      }
-    });
-  });
-
-  return columns;
-};
 
 const Configure: FC<ConfigureProps> = ({ sheets }) => {
   //settings
